+++ packages/gamut/src/AccordionArea/index.tsx
@@ -1,3 +1,4 @@
+import styled from '@emotion/styled';
 import { motion } from 'framer-motion';
 import React, { useLayoutEffect, useState } from 'react';
 
@@ -22,8 +23,8 @@ export type AccordionAreaProps = {
 const transitionDuration = 0.2;
 
 const variants = {
-  expanded: { height: '100%' },
-  folded: { height: 0 },
+  expanded: { height: 'auto' },
+  folded: { height: '0' },
 };
 
 export const AccordionArea: React.FC<AccordionAreaProps> = ({
@@ -44,17 +45,21 @@ export const AccordionArea: React.FC<AccordionAreaProps> = ({
   }, [expanded]);
 
   return (
-    <Box className={className} overflow="hidden">
+    <Box className={className}>
       {top}
-      <motion.div
+      <StyledAccordionBody
         aria-expanded={expanded}
-        initial={false}
+        initial={expanded ? 'expanded' : 'folded'}
         animate={expanded ? 'expanded' : 'folded'}
         variants={variants}
         transition={{ duration: transitionDuration, ease: 'easeInOut' }}
       >
         {(expanded || delayExpanded) && children}
-      </motion.div>
+      </StyledAccordionBody>
     </Box>
   );
 };
+
+const StyledAccordionBody = styled(motion.div)`
+  overflow: hidden;
+`;